Build lua-bcrypt.dll/.so too, fixes the library on Windows
Windows has a system DLL called bcrypt.dll. Even if we fix lua-bcrypt to
manually LoadLibrary/GetProcAddress BCryptGenRandom this is still a
landmine if the application otherwise depends on bcrypt.dll, so we have
to rename the library.

Since lua-bcrypt never worked on Windows we have no backwards
compatibility to preserve, so just ship lua-bcrypt.dll. Everyone else
gets bcrypt.so and lua-bcrypt.so, which are the same thing.

Future users should use `require("lua-bcrypt")`, but I will not break
`require("bcrypt")`.

Fixes #22, CC #23